# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Invoice Automation System API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ExtractedInvoiceData {
  vendor?: string | null;
  tax_id?: string | null;
  invoice_number?: string | null;
  date?: string | null;
  subtotal?: number | null;
  vat?: number | null;
  total?: number | null;
  currency?: string | null;
}

export interface ProcessInvoiceRequest {
  /** Path to the invoice file on disk */
  filePath: string;
  extracted: ExtractedInvoiceData;
  extractionConfidence?: number | null;
}

export interface ProcessInvoiceResponse {
  invoiceId: string;
  vendorId?: string | null;
  canonicalVendorName?: string | null;
  vendorMatchMethod?: string | null;
  duplicateStatus: string;
  duplicateOfInvoiceId?: string | null;
  confidence: number;
  status: string;
}

export interface InvoiceRow {
  id: string;
  rawVendorName?: string | null;
  normalizedVendorName?: string | null;
  canonicalVendorName?: string | null;
  taxId?: string | null;
  invoiceNumber?: string | null;
  invoiceDate?: string | null;
  subtotal?: string | null;
  vat?: string | null;
  total?: string | null;
  currency: string;
  filePath: string;
  duplicateStatus: string;
  duplicateOfInvoiceId?: string | null;
  status: string;
  extractionConfidence?: string | null;
  vendorId?: string | null;
  createdAt: string;
}

export interface MergeAliasRequest {
  aliasName: string;
  targetVendorId: string;
}

export type VendorWithAliasesAliasesItem = {
  id: string;
  aliasName: string;
  normalizedAlias: string;
};

export interface VendorWithAliases {
  id: string;
  canonicalName: string;
  taxId?: string | null;
  aliases: VendorWithAliasesAliasesItem[];
}

export type ListInvoicesParams = {
  status?: string;
  duplicate_status?: string;
};