# lib/api-zod/src/generated/types/extractedInvoiceData.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Invoice Automation System API
 * OpenAPI spec version: 0.1.0
 */

export interface ExtractedInvoiceData {
  vendor?: string | null;
  tax_id?: string | null;
  invoice_number?: string | null;
  date?: string | null;
  subtotal?: number | null;
  vat?: number | null;
  total?: number | null;
  currency?: string | null;
}