# lib/api-zod/src/generated/types/invoiceRow.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Invoice Automation System API
 * OpenAPI spec version: 0.1.0
 */

export interface InvoiceRow {
  id: string;
  rawVendorName?: string | null;
  normalizedVendorName?: string | null;
  canonicalVendorName?: string | null;
  taxId?: string | null;
  invoiceNumber?: string | null;
  invoiceDate?: string | null;
  subtotal?: string | null;
  vat?: string | null;
  total?: string | null;
  currency: string;
  filePath: string;
  duplicateStatus: string;
  duplicateOfInvoiceId?: string | null;
  status: string;
  extractionConfidence?: string | null;
  vendorId?: string | null;
  createdAt: string;
}