# lib/api-zod/src/generated/types/mergeAliasRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Invoice Automation System API
 * OpenAPI spec version: 0.1.0
 */

export interface MergeAliasRequest {
  aliasName: string;
  targetVendorId: string;
}