# lib/api-zod/src/generated/types/processInvoiceRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Invoice Automation System API
 * OpenAPI spec version: 0.1.0
 */
import type { ExtractedInvoiceData } from "./extractedInvoiceData";

export interface ProcessInvoiceRequest {
  /** Path to the invoice file on disk */
  filePath: string;
  extracted: ExtractedInvoiceData;
  extractionConfidence?: number | null;
}