# lib/api-zod/src/generated/types/processInvoiceResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Invoice Automation System API
 * OpenAPI spec version: 0.1.0
 */

export interface ProcessInvoiceResponse {
  invoiceId: string;
  vendorId?: string | null;
  canonicalVendorName?: string | null;
  vendorMatchMethod?: string | null;
  duplicateStatus: string;
  duplicateOfInvoiceId?: string | null;
  confidence: number;
  status: string;
}