# lib/api-zod/src/generated/types/vendorWithAliases.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Invoice Automation System API
 * OpenAPI spec version: 0.1.0
 */
import type { VendorWithAliasesAliasesItem } from "./vendorWithAliasesAliasesItem";

export interface VendorWithAliases {
  id: string;
  canonicalName: string;
  taxId?: string | null;
  aliases: VendorWithAliasesAliasesItem[];
}